Major restructuring Moved some code around
